+++ src/pages/LogIn.js
@@ -1,86 +1,41 @@
-import { Box, Center } from "@chakra-ui/react";
-import {Form, FormLayout, Field, SubmitButton } from "@saas-ui/react";
-import { supabase } from "@/utils/supabase";
-import {  useEffect } from "react";
-import { useRouter } from 'next/router';
+//   nextjs/reract imports
+import { useEffect, useState } from "react";
+import { useRouter } from "next/router";
+import { Link } from "next/link";
+
+// chakraUI/saas-ui imports
+import { Box, Center, Stack } from "@chakra-ui/react";
+import { Form, FormLayout, Field, SubmitButton } from "@saas-ui/react";
+
+// saas ui's auth imports
+import { Auth, PasswordForm } from "@saas-ui/react";
 
 export default function LogIn() {
-
   const router = useRouter();
+  const [errorMessage, setErrorMessage] = useState(null);
 
-  useEffect (() => {
-
-  supabase.auth.onAuthStateChange((event, session) => {
-    
-    if (session){
-        router.push('/')
-    }
-  })
-  
-  },[router]) 
-  const handleSubmit  = async({email, password}) => {
-    try {
-        const { data, error } = await supabase.auth.signInWithPassword({
-            email,
-            password,
-          })
-          if (data && data.user){
-            router.push('/')
-          }
-          
-    } catch (error) {
-        console.log(error)
-    }
-    
-
-  };
- 
-
-return(
-  
-  <Box padding="8" position='relative'>
-  <Center>
-  <Form
-    onSubmit={handleSubmit}
-  >
-    <FormLayout>
-   
-     
-      <Field
-        backgroundColor={"white"}
-        textColor={"black"}
-        name="email"
-        
-        label="Email"
-        type="email"
-        rules={{ required: true }}
-        width="20.5em"
-       
-      />
-      <Field
-        fontFamily={"mono"}
-        backgroundColor={"white"}
-        textColor={"black"}
-        type="password"
-        name="password"
-        label=" Password"
-        width="18em"
-        rules={{ required: true }}
-       
-      />
-     
-      <SubmitButton 
-      marginBlockStart="10px" disableIfInvalid>
-        Log In
-      </SubmitButton>
-    </FormLayout>
-  </Form>
-  </Center>
-</Box>
-
-
+  return (
+    <>
+      <Box padding="8" position="relative">
+        <Center h="90vh" w="100%">
+          <Stack maxW="400px">
+            <Auth
+              type="password"
+              signupLink={<a href="/signup">Sign up</a>}
+              onChange={() => {
+                setErrorMessage(null);
+              }}
+              onSuccess={() => {
+                router.push("/");
+              }}
+              onError={(error) => {
+                setErrorMessage(error.message);
+              }}
+            />
+            {errorMessage && <p className="highlight"> {errorMessage}</p>}
+          </Stack>
+        </Center>
+      </Box>
+    </>
   );
 }
-
-
-